perf(alembic): add integration records
