+++ alembic/versions/9f8e5aec59ea_adds_integration_providers.py
@@ -1,0 +1,294 @@
+"""adds integration providers
+
+Revision ID: 9f8e5aec59ea
+Revises: eb96f9b82cc1
+Create Date: 2025-05-27 07:52:57.549364
+
+"""
+
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+# revision identifiers, used by Alembic.
+revision = "9f8e5aec59ea"
+down_revision = "eb96f9b82cc1"
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    op.execute("create schema integration;")
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "integration_access",
+        sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
+        sa.Column("created_by", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("created_at", sa.DateTime(), nullable=True),
+        sa.Column("organization_id", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("integration_types", sa.ARRAY(sa.String()), nullable=True),
+        sa.ForeignKeyConstraint(["created_by"], ["user.id"], ondelete="SET NULL"),
+        sa.ForeignKeyConstraint(
+            ["organization_id"], ["organization.id"], ondelete="CASCADE"
+        ),
+        sa.PrimaryKeyConstraint("id"),
+        schema="cognition",
+    )
+    op.create_index(
+        op.f("ix_cognition_integration_access_organization_id"),
+        "integration_access",
+        ["organization_id"],
+        unique=False,
+        schema="cognition",
+    )
+    op.create_table(
+        "integration",
+        sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
+        sa.Column("organization_id", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("project_id", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("created_by", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("created_at", sa.DateTime(), nullable=True),
+        sa.Column("started_at", sa.DateTime(), nullable=True),
+        sa.Column("finished_at", sa.DateTime(), nullable=True),
+        sa.Column("name", sa.String(), nullable=True),
+        sa.Column("description", sa.String(), nullable=True),
+        sa.Column("tokenizer", sa.String(), nullable=True),
+        sa.Column("state", sa.String(), nullable=True),
+        sa.Column("type", sa.String(), nullable=True),
+        sa.Column("config", sa.JSON(), nullable=True),
+        sa.Column("llm_config", sa.JSON(), nullable=True),
+        sa.Column("error_message", sa.String(), nullable=True),
+        sa.ForeignKeyConstraint(["created_by"], ["user.id"], ondelete="SET NULL"),
+        sa.ForeignKeyConstraint(
+            ["organization_id"], ["organization.id"], ondelete="CASCADE"
+        ),
+        sa.ForeignKeyConstraint(["project_id"], ["project.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+        schema="cognition",
+    )
+    op.create_index(
+        op.f("ix_cognition_integration_organization_id"),
+        "integration",
+        ["organization_id"],
+        unique=False,
+        schema="cognition",
+    )
+    op.create_index(
+        op.f("ix_cognition_integration_project_id"),
+        "integration",
+        ["project_id"],
+        unique=False,
+        schema="cognition",
+    )
+    op.create_table(
+        "github_file",
+        sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
+        sa.Column("created_by", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("updated_by", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("created_at", sa.DateTime(), nullable=True),
+        sa.Column("updated_at", sa.DateTime(), nullable=True),
+        sa.Column("integration_id", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("running_id", sa.Integer(), nullable=True),
+        sa.Column("source", sa.String(), nullable=True),
+        sa.Column("path", sa.String(), nullable=True),
+        sa.Column("sha", sa.String(), nullable=True),
+        sa.Column("delta_criteria", sa.JSON(), nullable=True),
+        sa.Column("minio_file_name", sa.String(), nullable=True),
+        sa.ForeignKeyConstraint(["created_by"], ["user.id"], ondelete="SET NULL"),
+        sa.ForeignKeyConstraint(
+            ["integration_id"], ["cognition.integration.id"], ondelete="CASCADE"
+        ),
+        sa.ForeignKeyConstraint(["updated_by"], ["user.id"], ondelete="SET NULL"),
+        sa.PrimaryKeyConstraint("id"),
+        schema="integration",
+    )
+    op.create_index(
+        op.f("ix_integration_github_file_integration_id"),
+        "github_file",
+        ["integration_id"],
+        unique=False,
+        schema="integration",
+    )
+    op.create_index(
+        op.f("ix_integration_github_file_running_id"),
+        "github_file",
+        ["running_id"],
+        unique=False,
+        schema="integration",
+    )
+    op.create_index(
+        op.f("ix_integration_github_file_source"),
+        "github_file",
+        ["source"],
+        unique=False,
+        schema="integration",
+    )
+    op.create_table(
+        "github_issue",
+        sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
+        sa.Column("created_by", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("updated_by", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("created_at", sa.DateTime(), nullable=True),
+        sa.Column("updated_at", sa.DateTime(), nullable=True),
+        sa.Column("integration_id", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("running_id", sa.Integer(), nullable=True),
+        sa.Column("source", sa.String(), nullable=True),
+        sa.Column("url", sa.String(), nullable=True),
+        sa.Column("state", sa.String(), nullable=True),
+        sa.Column("assignee", sa.String(), nullable=True),
+        sa.Column("milestone", sa.String(), nullable=True),
+        sa.Column("number", sa.Integer(), nullable=True),
+        sa.Column("delta_criteria", sa.JSON(), nullable=True),
+        sa.Column("minio_file_name", sa.String(), nullable=True),
+        sa.ForeignKeyConstraint(["created_by"], ["user.id"], ondelete="SET NULL"),
+        sa.ForeignKeyConstraint(
+            ["integration_id"], ["cognition.integration.id"], ondelete="CASCADE"
+        ),
+        sa.ForeignKeyConstraint(["updated_by"], ["user.id"], ondelete="SET NULL"),
+        sa.PrimaryKeyConstraint("id"),
+        schema="integration",
+    )
+    op.create_index(
+        op.f("ix_integration_github_issue_integration_id"),
+        "github_issue",
+        ["integration_id"],
+        unique=False,
+        schema="integration",
+    )
+    op.create_index(
+        op.f("ix_integration_github_issue_running_id"),
+        "github_issue",
+        ["running_id"],
+        unique=False,
+        schema="integration",
+    )
+    op.create_index(
+        op.f("ix_integration_github_issue_source"),
+        "github_issue",
+        ["source"],
+        unique=False,
+        schema="integration",
+    )
+    op.create_table(
+        "pdf",
+        sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
+        sa.Column("created_by", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("updated_by", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("created_at", sa.DateTime(), nullable=True),
+        sa.Column("updated_at", sa.DateTime(), nullable=True),
+        sa.Column("integration_id", postgresql.UUID(as_uuid=True), nullable=True),
+        sa.Column("running_id", sa.Integer(), nullable=True),
+        sa.Column("source", sa.String(), nullable=True),
+        sa.Column("file_path", sa.String(), nullable=True),
+        sa.Column("page", sa.Integer(), nullable=True),
+        sa.Column("total_pages", sa.Integer(), nullable=True),
+        sa.Column("title", sa.String(), nullable=True),
+        sa.Column("delta_criteria", sa.JSON(), nullable=True),
+        sa.Column("minio_file_name", sa.String(), nullable=True),
+        sa.ForeignKeyConstraint(["created_by"], ["user.id"], ondelete="SET NULL"),
+        sa.ForeignKeyConstraint(
+            ["integration_id"], ["cognition.integration.id"], ondelete="CASCADE"
+        ),
+        sa.ForeignKeyConstraint(["updated_by"], ["user.id"], ondelete="SET NULL"),
+        sa.PrimaryKeyConstraint("id"),
+        schema="integration",
+    )
+    op.create_index(
+        op.f("ix_integration_pdf_integration_id"),
+        "pdf",
+        ["integration_id"],
+        unique=False,
+        schema="integration",
+    )
+    op.create_index(
+        op.f("ix_integration_pdf_running_id"),
+        "pdf",
+        ["running_id"],
+        unique=False,
+        schema="integration",
+    )
+    op.create_index(
+        op.f("ix_integration_pdf_source"),
+        "pdf",
+        ["source"],
+        unique=False,
+        schema="integration",
+    )
+    op.add_column(
+        "embedding",
+        sa.Column("delta_full_recalculation_threshold", sa.Float(), nullable=True),
+    )
+    op.add_column(
+        "embedding",
+        sa.Column("current_delta_record_count", sa.Integer(), nullable=True),
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column("embedding", "current_delta_record_count")
+    op.drop_column("embedding", "delta_full_recalculation_threshold")
+    op.drop_index(
+        op.f("ix_integration_pdf_source"), table_name="pdf", schema="integration"
+    )
+    op.drop_index(
+        op.f("ix_integration_pdf_running_id"), table_name="pdf", schema="integration"
+    )
+    op.drop_index(
+        op.f("ix_integration_pdf_integration_id"),
+        table_name="pdf",
+        schema="integration",
+    )
+    op.drop_table("pdf", schema="integration")
+    op.drop_index(
+        op.f("ix_integration_github_issue_source"),
+        table_name="github_issue",
+        schema="integration",
+    )
+    op.drop_index(
+        op.f("ix_integration_github_issue_running_id"),
+        table_name="github_issue",
+        schema="integration",
+    )
+    op.drop_index(
+        op.f("ix_integration_github_issue_integration_id"),
+        table_name="github_issue",
+        schema="integration",
+    )
+    op.drop_table("github_issue", schema="integration")
+    op.drop_index(
+        op.f("ix_integration_github_file_source"),
+        table_name="github_file",
+        schema="integration",
+    )
+    op.drop_index(
+        op.f("ix_integration_github_file_running_id"),
+        table_name="github_file",
+        schema="integration",
+    )
+    op.drop_index(
+        op.f("ix_integration_github_file_integration_id"),
+        table_name="github_file",
+        schema="integration",
+    )
+    op.drop_table("github_file", schema="integration")
+    op.drop_index(
+        op.f("ix_cognition_integration_project_id"),
+        table_name="integration",
+        schema="cognition",
+    )
+    op.drop_index(
+        op.f("ix_cognition_integration_organization_id"),
+        table_name="integration",
+        schema="cognition",
+    )
+    op.drop_table("integration", schema="cognition")
+    op.drop_index(
+        op.f("ix_cognition_integration_access_organization_id"),
+        table_name="integration_access",
+        schema="cognition",
+    )
+    op.drop_table("integration_access", schema="cognition")
+    # ### end Alembic commands ###
+    op.execute("drop schema integration;")
