--- alembic/versions/cee210baa0ea_adds_integration_providers.py
@@ -1,115 +1,0 @@
-"""adds integration providers
-
-Revision ID: cee210baa0ea
-Revises: eb96f9b82cc1
-Create Date: 2025-05-26 15:18:32.868727
-
-"""
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = "cee210baa0ea"
-down_revision = "eb96f9b82cc1"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "integration_access",
-        sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-        sa.Column("created_by", postgresql.UUID(as_uuid=True), nullable=True),
-        sa.Column("created_at", sa.DateTime(), nullable=True),
-        sa.Column("organization_id", postgresql.UUID(as_uuid=True), nullable=True),
-        sa.Column("integration_types", sa.ARRAY(sa.String()), nullable=True),
-        sa.ForeignKeyConstraint(["created_by"], ["user.id"], ondelete="SET NULL"),
-        sa.ForeignKeyConstraint(
-            ["organization_id"], ["organization.id"], ondelete="CASCADE"
-        ),
-        sa.PrimaryKeyConstraint("id"),
-        schema="cognition",
-    )
-    op.create_index(
-        op.f("ix_cognition_integration_access_organization_id"),
-        "integration_access",
-        ["organization_id"],
-        unique=False,
-        schema="cognition",
-    )
-    op.create_table(
-        "integration",
-        sa.Column("id", postgresql.UUID(as_uuid=True), nullable=False),
-        sa.Column("organization_id", postgresql.UUID(as_uuid=True), nullable=True),
-        sa.Column("project_id", postgresql.UUID(as_uuid=True), nullable=True),
-        sa.Column("created_by", postgresql.UUID(as_uuid=True), nullable=True),
-        sa.Column("created_at", sa.DateTime(), nullable=True),
-        sa.Column("started_at", sa.DateTime(), nullable=True),
-        sa.Column("finished_at", sa.DateTime(), nullable=True),
-        sa.Column("name", sa.String(), nullable=True),
-        sa.Column("description", sa.String(), nullable=True),
-        sa.Column("tokenizer", sa.String(), nullable=True),
-        sa.Column("state", sa.String(), nullable=True),
-        sa.Column("type", sa.String(), nullable=True),
-        sa.Column("config", sa.JSON(), nullable=True),
-        sa.Column("llm_config", sa.JSON(), nullable=True),
-        sa.Column("error_message", sa.String(), nullable=True),
-        sa.Column("extract_history", sa.JSON(), nullable=True),
-        sa.ForeignKeyConstraint(["created_by"], ["user.id"], ondelete="SET NULL"),
-        sa.ForeignKeyConstraint(
-            ["organization_id"], ["organization.id"], ondelete="CASCADE"
-        ),
-        sa.ForeignKeyConstraint(["project_id"], ["project.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-        schema="cognition",
-    )
-    op.create_index(
-        op.f("ix_cognition_integration_organization_id"),
-        "integration",
-        ["organization_id"],
-        unique=False,
-        schema="cognition",
-    )
-    op.create_index(
-        op.f("ix_cognition_integration_project_id"),
-        "integration",
-        ["project_id"],
-        unique=False,
-        schema="cognition",
-    )
-    op.add_column(
-        "embedding",
-        sa.Column("delta_full_recalculation_threshold", sa.Float(), nullable=True),
-    )
-    op.add_column(
-        "embedding",
-        sa.Column("current_delta_record_count", sa.Integer(), nullable=True),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column("embedding", "current_delta_record_count")
-    op.drop_column("embedding", "delta_full_recalculation_threshold")
-    op.drop_index(
-        op.f("ix_cognition_integration_project_id"),
-        table_name="integration",
-        schema="cognition",
-    )
-    op.drop_index(
-        op.f("ix_cognition_integration_organization_id"),
-        table_name="integration",
-        schema="cognition",
-    )
-    op.drop_table("integration", schema="cognition")
-    op.drop_index(
-        op.f("ix_cognition_integration_access_organization_id"),
-        table_name="integration_access",
-        schema="cognition",
-    )
-    op.drop_table("integration_access", schema="cognition")
-    # ### end Alembic commands ###
